requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7
datasets==3.2.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==2.3.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.9.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.1
ollama==0.3.3
onnx==1.17.0
onnxruntime-gpu==1.20.1
openai==1.54.4
optimum==1.23.3
osc4py3==1.0.8
packaging==24.2
pandas==2.2.3
propcache==0.2.1
protobuf==5.29.2
pyarrow==18.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-engineio==4.6.1
python-socketio==5.8.0
python-statemachine==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors>=0.4.5
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.0
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3